requirements_splade.txt
ransformers torch
git+https://github.com/naver/splade.git
faiss-gpu